requirements.txt
# AI Podcast Creator - Dependencies

# Core Framework
typer[all]==0.9.0
rich==13.7.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML Core
torch==2.1.0
torchaudio==2.1.0
transformers==4.35.0
diffusers==0.24.0
accelerate==0.25.0

# TTS Engines
TTS==0.22.0
# Optional commercial APIs (uncomment if using):
# elevenlabs==0.2.27
# azure-cognitiveservices-speech==1.34.0

# Audio Processing
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1
scipy==1.11.4
numpy==1.24.3

# Music Generation
audiocraft==1.3.0

# Video Processing
moviepy==1.0.3
opencv-python==4.8.1.78
pillow==10.1.0
imageio==2.33.0
imageio-ffmpeg==0.4.9

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Task Queue (Optional)
celery==5.3.4
redis==5.0.1

# API Framework (Optional)
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2
tenacity==8.2.3
loguru==0.7.2
tqdm==4.66.1
pyyaml==6.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

